backend/requirements.txt
Flask==1.1.2
pandas==1.3.4
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
psycopg2==2.8.6
pylint-flask-sqlalchemy==0.2.0
pylint-flask==0.6
pylint-plugin-utils==0.6